requirements.txt
langchain
langchain-community
langchain-core
langchain-openai>=0.1.0
faiss-cpu
sentence-transformers>=2.2.0
huggingface-hub
pypdf
requests
pydantic
pdfplumber==0.11.0
pytesseract==0.3.10
pdf2image==1.16.3
fastapi
uvicorn
python-dotenv>=1.0.0
